🛠️ moveディレクトリ内に、移動と衝突の関数を入れている
